+++ Sensors@claudiux/files/Sensors@claudiux/scripts/make_smartctl_usable_by_sudoers.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+[[ $UID -eq 0 ]] || exit 1
+echo "%sudo ALL = NOPASSWD: /usr/sbin/smartctl" | sudo tee "/etc/sudoers.d/smartctl"
+
+exit 0