+++ radio@driglu4it/src/index.ts
@@ -1,8 +1,9 @@
 import { initConfig } from './services/Config';
 import { initMpvHandler } from './services/mpv/MpvHandler';
 import { initPolyfills } from './polyfill';
-import {  getRadioAppletContainer } from './ui/RadioApplet/RadioAppletContainer';
-
+import { getRadioAppletContainer } from './ui/RadioApplet/RadioAppletContainer';
+import { APPLET_CACHE_DIR_PATH } from './consts';
+const { new_for_path } = imports.gi.Gio.File
 declare global {
     // added during build (see webpack.config.js)
     interface Meta {
@@ -12,9 +13,14 @@ declare global {
         locationLabel: imports.ui.appletManager.LocationLabel
     }
 }
-
  
 const onAppletMovedCallbacks: Array<() => void> = []
+
+const createCacheDir = () => {
+    const dir = new_for_path(APPLET_CACHE_DIR_PATH);
+    if (!dir.query_exists(null)) dir.make_directory_with_parents(null)
+}
+
 
 export const addOnAppletMovedCallback = (cb: () => void) => {
     onAppletMovedCallbacks.push(cb)
@@ -22,6 +28,8 @@ export const addOnAppletMovedCallback = (cb: () => void) => {
 
 // The function defintion must use the word "function" (not const!) as otherwilse the error: "radioApplet.main is not a constructor" is thrown
 export function main(): imports.ui.applet.Applet {
+
+    createCacheDir();
 
     // order must be retained!
     initPolyfills()