+++ radio@driglu4it/src/services/mpv/CheckInstallation.ts
@@ -1,13 +1,10 @@
 import { spawnCommandLinePromise } from "../../functions/promiseHelpers";
-
 import { MPRIS_PLUGIN_PATH, MPRIS_PLUGIN_URL } from '../../consts'
 import { notify } from "../../lib/notify";
-
-
 const { find_program_in_path, file_test, FileTest } = imports.gi.GLib;
 
-export async function installMpvWithMpris() {
 
+export async function installMpvWithMpris() {
     const mprisPluginDownloaded = checkMprisPluginDownloaded()
     const mpvInstalled = checkMpvInstalled()