+++ system-controls@rcalixte/CHANGELOG.md
@@ -1,4 +1,8 @@
 
+### 1.1.1
+ * Changed icon for 'Restart Cinnamon' menu item from `system-run` to `cinnamon-symbolic`
+ * Improved packaged images
+
 ### 1.1
  * Added proper translation support
  * Created dedicated changelog file